LDT MAV DS 4.3, Part Changes
LDT parts were updated.  Textures were replaced for the HAB tanks & MAV
aeroshell.  Node positions were updated for the MAV aeroshell, and the
engine mountings moved to a separate part.  The  MAV DS was updated to
version 4.3b, which reflected the change in the aeroshell and also had
minor optimization changes.